Break out proposal module from db
